Favorite Unit Selection Scene
A scene that loads the fav unit of the user and you can select other to change it (not implemented yet).
Sometimes firebase have errors on loading the fav unit, but no others variables like the username. (maybe have something to see with the loading errors we got on the presentation)